mapstructure 테스트중

+++ src/main/java/com/example/spring_blog/service/CommentService.java
@@ -31,7 +31,7 @@ public class CommentService {
         return Optional.ofNullable(commentRepository.save(comment));
     }
 
-    public List<Comment> findAllByPostId(Integer postId) {
+    public List<Comment> findAllByPostId(Long postId) {
         return commentRepository.findAllByPost(Post.builder().id(postId).build());
     }
